refactor: split actix-http test helpers
