+++ actix-http/src/test/mod.rs
@@ -1,0 +1,11 @@
+//! Various testing helpers for use in internal and app tests.
+
+mod test_buffer;
+mod test_request;
+mod test_seq_buffer;
+
+pub use self::{
+    test_buffer::TestBuffer,
+    test_request::TestRequest,
+    test_seq_buffer::{TestSeqBuffer, TestSeqInner},
+};
